requirements.txt
blinker==1.7.0
click==8.1.7
confluent-kafka==2.3.0
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kafka-python==2.0.2
MarkupSafe==2.1.3
nltk==3.8.1
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.10.3
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
Werkzeug==3.0.1
